clippy: take the workspace to zero warnings
`cargo clippy --workspace --all-targets` stood at 67 warnings, which is enough
noise to hide a real one. It is now silent, split between substantive fixes and
two lints the project declines on the record.

Fixed:
  - apps.rs           CreateProcessW takes STARTUPINFOW by const pointer; the
                      `&mut` was mine, from the uninstall rewrite
  - settings.rs       `.map(|(v, m)| (v, m))` was the identity function
  - trayoverflow.rs   `.min(COLS).max(1)` spelled as `.clamp(1, COLS)`
  - startmenu.rs      hit test as a range `contains`, same bounds
  - taskbar.rs        ditto for the Start-button hit test
  - glide/app.rs      `!x.is_some_and(|e| a == b)` → `x.is_none_or(|e| a != b)`
  - shellmenu.rs ×2   a `let…else { return None }` inside an Option closure is
                      just `?`
  - osd.rs, toasts.rs `HTTRANSPARENT as i32 as isize` — HTTRANSPARENT is i32
  - flyout.rs         the volume panel's section tuple named as DeviceSection
  - taskbar.rs        vec_box allowed with a reason: Secondary hands its own
                      address to GWLP_USERDATA, so following the lint's advice
                      would leave the wndproc dereferencing moved memory
  - taskmgr.rs        module doc reflowed off a line starting with `+`, which
                      rustdoc was reading as a list item

Allowed at the workspace level, with the reasoning in Cargo.toml:
  - collapsible_if (49 sites) — since the 2024 edition this asks for let-chains.
    The win32 handlers are written as nested guards throughout; converting a
    subset would leave the crate in two styles, and `clippy --fix` cannot
    reindent the bodies it rewrites, so the mechanical route produces malformed
    blocks. Not worth reformatting the codebase over a style lint.
  - too_many_arguments (4 sites) — the D2D and egui painters take geometry and
    palette positionally.

Every rewrite above is an equivalence, not a behaviour change. Build clean,
`cargo test --workspace` 6 passed, and the Task Manager still launches and
renders.

+++ crates/glide-shell/src/osd.rs
@@ -512,7 +512,7 @@ extern "system" fn osd_wndproc(hwnd: HWND, msg: u32, wparam: WPARAM, lparam: LPA
             }
             WM_ERASEBKGND => LRESULT(1),
             // Display-only: everything falls through to whatever is beneath.
-            WM_NCHITTEST => LRESULT(HTTRANSPARENT as i32 as isize),
+            WM_NCHITTEST => LRESULT(HTTRANSPARENT as isize),
             WM_APP_VOL => {
                 let muted = wparam.0 != 0;
                 let vol = lparam.0 as f32 / 1000.0;
